backend/requirements.txt
fastapi
uvicorn
pillow
rembg
diffusers
torch
